Baranyi and Roberts full growth model

BaranyiFM function to fit the Baranyi & Roberts full growth model to a complete microbial growth curve. Returns the model parameters estimated according to data collected in microbial growth experiments.

Usage

BaranyiFM(t, Y0, Ymax, MUmax, lag)

Arguments

t

is a numeric vector indicating the time of the experiment

Y0

is the natural logarithm of the initial microbial concentration (ln(N0)) at time=0

Ymax

is the natural logarithm of the maximum concentration (ln(Nmax)) reached by the microorganism

MUmax

is the maximum specific growth rate given in time units

lag

is the duration of the lag phase in time units

Value

A numeric vector with the fitted values

Details

The model's inputs are:

t: time, assuming time zero as the beginning of the experiment.

Y(t): the natural logarithm of the microbial concentration (ln(N(t)) measured at time t.

Users should make sure that the microbial concentration input is entered in natural logarithm, Y(t) = ln(N(t)).

References

Baranyi1994predmicror

Author

Vasco Cadavez vcadavez@ipb.pt and Ursula Gonzales-Barron ubarron@ipb.pt

Examples

library(gslnls)
data(growthfull)
initial_values <- list(Y0 = -0.1, Ymax = 22, MUmax = 1.7, lag = 5)
fit <- gsl_nls(lnN ~ BaranyiFM(Time, Y0, Ymax, MUmax, lag),
  data = growthfull,
  start = initial_values
)
summary(fit)
#> 
#> Formula: lnN ~ BaranyiFM(Time, Y0, Ymax, MUmax, lag)
#> 
#> Parameters:
#>       Estimate Std. Error t value Pr(>|t|)    
#> Y0    -0.01043    0.27279  -0.038     0.97    
#> Ymax  21.13052    0.22783  92.748 9.98e-15 ***
#> MUmax  1.86815    0.06507  28.712 3.67e-10 ***
#> lag    5.08623    0.28282  17.984 2.31e-08 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.4525 on 9 degrees of freedom
#> 
#> Number of iterations to convergence: 8 
#> Achieved convergence tolerance: 1.776e-15
#>