Cardinal model for growth inhibitors

CMInh function to fit the growth inhibitors cardinal model (Rosso et al, 1993). Returns the model parameters estimated according to data collected in microbial growth experiments.

Usage

CMInh(x, MIC, MUopt, alpha)

Arguments

x

is a numeric vector indicating the inhibitor concentration of the experiment

MIC

is the minimum inhibitory concentration (mM or %, accordingly)

MUopt

is the optimum growth rate

alpha

is the shape parameter of the curve (alpha = 1 the shape is linear; alpha > 1 the shape is downward concave; and alpha < 1 the shape is upward concave)

Value

A numeric vector with the fitted values

Details

The model's inputs are:

x: growth inhibitor concentration

sqrtGR: the square root of the growth rate ($time^-1$)

Users should make sure that the growth rate input is entered after a square root transformation, $sqrGR = sqrt(GR)$.

References

Rosso1995predmicror

Author

Vasco Cadavez vcadavez@ipb.pt and Ursula Gonzales-Barron ubarron@ipb.pt

Examples

library(gslnls)
data(inh)
initial_values <- list(MIC = 0.89, MUopt = 1.0, alpha = 1)
fit <- gsl_nls(sqrtGR ~ CMInh(Conce, MIC, MUopt, alpha),
  data = inh,
  start = initial_values
)
summary(fit)
#> 
#> Formula: sqrtGR ~ CMInh(Conce, MIC, MUopt, alpha)
#> 
#> Parameters:
#>       Estimate Std. Error t value Pr(>|t|)  
#> MIC     3.7677     2.2659   1.663   0.1572  
#> MUopt   0.7416     0.3426   2.165   0.0827 .
#> alpha   0.6732     0.8736   0.771   0.4757  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.2011 on 5 degrees of freedom
#> 
#> Number of iterations to convergence: 10 
#> Achieved convergence tolerance: 2.929e-13
#>