Richards no lag growth model

RichardsNLM function to fit the Richards no lag growth model to an incomplete microbial growth curve. Returns the model parameters estimated according to data collected in microbial growth experiments.

Usage

RichardsNLM(t, Y0, Ymax, MUmax)

Arguments

t

is a numeric vector indicating the time of the experiment

Y0

is the natural logarithm of the initial microbial concentration (ln(N0)) at time=0

Ymax

is the natural logarithm of the maximum concentration (ln(Nmax)) reached by the microorganism

MUmax

is the maximum specific growth rate given in time units

Value

An object of nls class

Details

Model's inputs are:

t: time, assuming time zero as the beginning of the experiment.

Y(t): the natural logarithm of the microbial concentration (ln(N(t)) measured at time t.

Users should make sure that the microbial concentration input is entered in natural logarithm, Y(t) = ln(N(t)).

References

Richards JF (1959). “A flexible growth function for empirical use.” J Exp Bot, 1(10), 290-310.

Author

Vasco Cadavez, vcadavez@ipb.pt and Ursula Gonzales-Barron, ubarron@ipb.pt

Examples

## Example: Richards no lag model
library(gslnls)
data(growthnolag)  # simulated data set.
initial_values = list(Y0=0, Ymax=22, MUmax=1.7)
## Fitting the function to the experimental data
fit <- gsl_nls(lnN ~ RichardsNLM(Time, Y0, Ymax, MUmax),
               data=growthnolag,
               start =  initial_values)
summary(fit)
#> 
#> Formula: lnN ~ RichardsNLM(Time, Y0, Ymax, MUmax)
#> 
#> Parameters:
#>       Estimate Std. Error t value Pr(>|t|)    
#> Y0     1.75827    0.36368   4.835  0.00189 ** 
#> Ymax  21.13234    0.24673  85.650 7.79e-12 ***
#> MUmax  1.85922    0.06619  28.091 1.86e-08 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.4896 on 7 degrees of freedom
#> 
#> Number of iterations to convergence: 5 
#> Achieved convergence tolerance: 1.326e-12
#>