Weibull inactivation modified model Mafart

WeibullMM inactivation model for microbial inactivation curve. Returns the model parameters estimated according to data collected in microbial inactivation experiments.

Usage

WeibullMM(x, Y0, Yres, sigma, alpha)

Arguments

x

is a numeric vector indicating the heating time under a constant temperature of the experiment

Y0

is the natural logarithm of the initial bacterial concentration (at time t=0)

Yres

is the natural logarithm of the residual bacterial concentration (at the end of the experiment)

sigma

represents the time of the first decimal reduction concentration for the part of the population not belonging to Yres

alpha

is the shape parameter and allows to catch the curve concavity or convexity

Value

A numeric vector with the fitted values

Details

The model's inputs are:

t: time, assuming time zero as the beginning of the experiment.

Y(t): the bacterial concentration (Y(t)) measured at time t.

Users should make sure to use the natural logarithm bacterial concentration (Y(t)) as input.

References

Albert I, Mafart P (2005). “A modified Weibull model for bacterial inactivation.” International Journal of Food Microbiology, 100, 197-211.

Author

Vasco Cadavez vcadavez@ipb.pt and Ursula Gonzales-Barron ubarron@ipb.pt

Examples

library(gslnls)
data(bixina)
initial_values <- list(Y0 = 6, Yres = 1, sigma = 10, alpha = 2)
fit <- gsl_nls(lnN ~ WeibullMM(Time, Y0, Yres, sigma, alpha),
 data = bixina,
 start = initial_values
)
summary(fit)
#> 
#> Formula: lnN ~ WeibullMM(Time, Y0, Yres, sigma, alpha)
#> 
#> Parameters:
#>         Estimate Std. Error t value Pr(>|t|)    
#> Y0     5.533e+00  4.138e-02  133.74  < 2e-16 ***
#> Yres  -1.480e+01  1.202e+07    0.00        1    
#> sigma  1.287e+01  4.222e-01   30.49 3.34e-14 ***
#> alpha  2.395e+00  1.996e-01   12.00 9.37e-09 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.1001 on 14 degrees of freedom
#> 
#> Number of iterations to convergence: 33 
#> Achieved convergence tolerance: 4.162e-11
#>