Fit a microbial inactivation model

fit_inactivation() validates the input data, builds the nonlinear model formula, fits it with gslnls::gsl_nls(), and returns a predmicror_fit object.

Usage

fit_inactivation(data, model, time, response = "logN", start, ...)

Arguments

data

A data frame containing the time and response variables.

model

A model name, either quoted or unquoted. See predmicror_models().

time

Column containing time values, either quoted or unquoted.

response

Column containing the response values, either quoted or unquoted. Defaults to "logN".

start

Named list of initial parameter values for the selected model.

...

Additional arguments passed to gslnls::gsl_nls().

Value

A predmicror_fit object with the fitted model and metadata.

Details

Inactivation models expect the response to be the base 10 logarithm of the microbial concentration, usually logN.

Examples

data(mafart2005Li11)
fit <- fit_inactivation(
  mafart2005Li11,
  model = "WeibullM",
  time = "Time",
  response = "logN",
  start = list(Y0 = 10, sigma = 3, alpha = 1)
)
coef(fit)
#>         Y0      sigma      alpha 
#> 10.6393880 13.1941847  0.7727164