List models available through the fitting wrappers

List models available through the fitting wrappers

Usage

predmicror_models(type = "all")

Arguments

type

Character string. One of "all", "growth", "inactivation", or "cardinal".

Value

A data frame with the model type, model name, expected response scale, and parameters that must be supplied in start.

Examples

predmicror_models()
#>            type        model response_scale                 parameters
#> 1        growth      HuangFM            lnN       Y0, Ymax, MUmax, lag
#> 2        growth    BaranyiFM            lnN       Y0, Ymax, MUmax, lag
#> 3        growth ZwieteringFM            lnN       Y0, Ymax, MUmax, lag
#> 4        growth      RossoFM            lnN       Y0, Ymax, MUmax, lag
#> 5        growth      HuangRM            lnN             Y0, MUmax, lag
#> 6        growth    BaranyiRM            lnN             Y0, MUmax, lag
#> 7        growth   BuchananRM            lnN             Y0, MUmax, lag
#> 8        growth     HuangNLM            lnN            Y0, Ymax, MUmax
#> 9        growth      FangNLM            lnN            Y0, Ymax, MUmax
#> 10       growth  RichardsNLM            lnN         Y0, Ymax, MUmax, m
#> 11 inactivation     WeibullM         log10N           Y0, sigma, alpha
#> 12 inactivation    WeibullMM         log10N     Y0, Yres, sigma, alpha
#> 13 inactivation    WeibullPH         log10N               Y0, k, alpha
#> 14 inactivation   GeeraerdST         log10N         Y0, Yres, kmax, Sl
#> 15 inactivation     HuangRGS         log10N                   Y0, k, M
#> 16     cardinal         CMTI         sqrtGR    Tmax, Tmin, MUopt, Topt
#> 17     cardinal         CMAW         sqrtGR        AWmin, MUopt, AWopt
#> 18     cardinal         CMPH         sqrtGR pHmax, pHmin, MUopt, pHopt
#> 19     cardinal        CMInh         sqrtGR          MIC, MUopt, alpha
predmicror_models("growth")
#>      type        model response_scale           parameters
#> 1  growth      HuangFM            lnN Y0, Ymax, MUmax, lag
#> 2  growth    BaranyiFM            lnN Y0, Ymax, MUmax, lag
#> 3  growth ZwieteringFM            lnN Y0, Ymax, MUmax, lag
#> 4  growth      RossoFM            lnN Y0, Ymax, MUmax, lag
#> 5  growth      HuangRM            lnN       Y0, MUmax, lag
#> 6  growth    BaranyiRM            lnN       Y0, MUmax, lag
#> 7  growth   BuchananRM            lnN       Y0, MUmax, lag
#> 8  growth     HuangNLM            lnN      Y0, Ymax, MUmax
#> 9  growth      FangNLM            lnN      Y0, Ymax, MUmax
#> 10 growth  RichardsNLM            lnN   Y0, Ymax, MUmax, m