Weibull inactivation modified model Mafart

WeibullMM inactivation model for microbial inactivation curve. Returns the model parameters estimated according to data collected in microbial inactivation experiments.

Usage

WeibullMM(x, Y0, Yres, sigma, alpha)

Arguments

x

is a numeric vector indicating the heating time under a constant temperature of the experiment

Y0

is the log10 of the initial bacterial concentration (at time t=0)

Yres

is the log10 of the residual bacterial concentration (at the end of the experiment)

sigma

represents the time of the first decimal reduction concentration for the part of the population not belonging to Yres

alpha

is the shape parameter and allows to catch the curve concavity or convexity

Value

An object of nls class with the fitted parameters of the model

Details

The model's inputs are:

t: time, assuming time zero as the beginning of the experiment.

Y(t): the bacterial concentration ($Y(t)$) measured at time t.

Users should make sure to use the base 10 logarithm bacterial concentration (Y(t)) as input.

References

Albert I, Mafart P (2005). “A modified Weibull model for bacterial inactivation.” International Journal of Food Microbiology, 100, 197-211.

Author

Vasco Cadavez vcadavez@ipb.pt and Ursula Gonzales-Barron ubarron@ipb.pt

Examples

library(gslnls)
data(bixina)